+++ src/plonemeeting/portal/core/utils.py
@@ -97,7 +97,6 @@ def create_faceted_folder(container, title, id=None):
         )
     else:
         folder = api.content.create(type="Folder", title=title, container=container)
-    api.content.transition(folder, to_state="published")
     subtyper = folder.restrictedTraverse("@@faceted_subtyper")
     subtyper.enable()
     set_constrain_types(folder, [])